chore: saving

+++ src/modules/property/services/create-property.service.ts
@@ -165,7 +165,7 @@ export class CreateProperty_Service {
               creditCard: creditCardData,
               isCreate: true,
               ownerActiveProperties,
-              session
+              session,
             }
             const planTransitionData = await this.createPlanTransition(
               createPlanTransition,
@@ -188,11 +188,10 @@ export class CreateProperty_Service {
               ownerPreviousPlan,
               ownerActiveProperties,
               deactivateProperties,
-              session
+              session,
             )
           } else {
-            // updatedOwner.adCredits = updatedOwner.adCredits - 1;
-            await this.handlePayment(
+            updatedOwner = await this.handlePayment(
               isPlanFree,
               selectedPlan,
               owner,
@@ -203,7 +202,7 @@ export class CreateProperty_Service {
               ownerPreviousPlan,
               ownerActiveProperties,
               deactivateProperties,
-              session
+              session,
             )
           }
         }
@@ -240,8 +239,8 @@ export class CreateProperty_Service {
               ownerActiveProperties,
               deactivateProperties,
               updatedOwner.paymentData,
-              session
-            );
+              session,
+            )
           }
         }
       }
@@ -586,10 +585,10 @@ export class CreateProperty_Service {
     userData: UserData,
     creditCardData: CreditCardData,
     cellPhone: string,
-    ownerPreviousPlan: string,
+    ownerPreviousPlan: string | IPlan,
     ownerProperties: any,
     propsToDeactivate: any,
-    session: ClientSession
+    session: ClientSession,
   ) {
     let cpfCnpj: string
     let expiry: string
@@ -702,8 +701,8 @@ export class CreateProperty_Service {
             ownerProperties,
             propsToDeactivate,
             updatedOwner.paymentData,
-            session
-          );
+            session,
+          )
         } else {
           throw new Error(`Falha ao gerar a cobrança: ${response.statusText}`)
         }
@@ -762,28 +761,6 @@ export class CreateProperty_Service {
                   `Falha ao atualizar a assinatura: ${response.statusText}`,
                 )
               }
-
-              const responseData = response.data
-              const updatedSubscriptionId = responseData.id
-              const newAdcredits = owner.adCredits - 1 + selectedPlan.commonAd
-              const newHighlightCredits =
-                owner.highlightCredits > 0
-                  ? owner.highlightCredits - 1 + selectedPlan.highlightAd
-                  : owner.highlightCredits
-
-              //Atualizar os créditos do usuário
-              // await this.ownerModel.updateOne(
-              //   { _id: owner._id },
-              //   {
-              //     $set: {
-              //       adCredits: newAdcredits,
-              //       highlightCredits: newHighlightCredits,
-              //       plan: selectedPlan._id,
-              //       'paymentData.subscriptionId': updatedSubscriptionId,
-              //     },
-              //   },
-              //   { session },
-              // )
             } else {
               throw new Error(
                 `O usuário não tem mais créditos disponíveis para anunciar.`,
@@ -846,15 +823,6 @@ export class CreateProperty_Service {
                 },
               },
             )
-
-            // Atualização dos créditos;
-            // if (typeof creditsLeft === 'number') {
-            //   updatedOwner.adCredits = creditsLeft
-            // } else {
-            //   updatedOwner.adCredits =
-            //     selectedPlan.commonAd - 1 - ownerActiveProperties
-            //   updatedOwner.plan = selectedPlan._id
-            // }
           } catch (error) {}
         } else {
           throw new NotFoundException('Assinatura não encontrada.')
@@ -908,14 +876,8 @@ export class CreateProperty_Service {
 
             // Condicional dos créditos para casos em que já havia um plano anterior ou não;
             if (!previousPlanData) {
-              // newAdCredits = selectedPlan.commonAd
               newHighlightCredits = selectedPlan.highlightAd
             } else if (ownerActualPlanString !== planIdString) {
-              // newAdCredits =
-              //   selectedPlan.price > previousPlanData.price
-              //     ? owner.adCredits + selectedPlan.commonAd
-              //     : owner.adCredits - selectedPlan.commonAd
-
               newHighlightCredits =
                 selectedPlan.price > previousPlanData.price
                   ? owner.highlightCredits + selectedPlan.highlightAd
@@ -940,43 +902,57 @@ export class CreateProperty_Service {
             }
           }
         } else {
-          if (owner.paymentData?.subscriptionId) {
-            // Cancelar a assinatura
-            const { data } = await axios.delete(
-              `${process.env.PAYMENT_URL}/payment/subscription/${owner?.paymentData?.subscriptionId}`,
-              {
-                method: 'DELETE',
-                headers: {
-                  'Content-Type': 'application/json',
-                  access_token: process.env.ASSAS_API_KEY || '',
+          if (
+            typeof ownerPreviousPlan !== 'string' &&
+            ownerActualPlanString !== ownerPreviousPlan?._id.toString()
+          ) {
+            if (owner.paymentData?.subscriptionId) {
+              // Cancelar a assinatura
+              const { data } = await axios.delete(
+                `${process.env.PAYMENT_URL}/payment/subscription/${owner?.paymentData?.subscriptionId}`,
+                {
+                  method: 'DELETE',
+                  headers: {
+                    'Content-Type': 'application/json',
+                    access_token: process.env.ASSAS_API_KEY || '',
+                  },
                 },
-              },
-            )
+              )
 
-            const success = data.deleted
+              const success = data.deleted
 
-            if (!success) {
-              throw new Error('Não foi possível remover a assinatura')
-            }
+              if (!success) {
+                throw new Error('Não foi possível remover a assinatura')
+              }
 
-            updatedOwner.paymentData.creditCardInfo = {
-              creditCardBrand: '',
-              creditCardNumber: '',
-              creditCardToken: '',
-            }
-            updatedOwner.paymentData.subscriptionId = ''
+              updatedOwner.paymentData.creditCardInfo = {
+                creditCardBrand: '',
+                creditCardNumber: '',
+                creditCardToken: '',
+              }
+              updatedOwner.paymentData.subscriptionId = ''
 
-            updatedOwner = {
-              ...updatedOwner,
-              adCredits: owner.adCredits - 1,
-            }
-          } else {
-            updatedOwner = {
-              ...updatedOwner,
-              adCredits: owner.adCredits - 1,
-              highlightCredits: 0,
+              updatedOwner = {
+                ...updatedOwner,
+                adCredits: owner.adCredits - 1,
+              }
+            } else {
+              updatedOwner = {
+                ...updatedOwner,
+                adCredits: owner.adCredits - 1,
+                highlightCredits: 0,
+              }
             }
           }
+          updatedOwner = await this.updateCredits(
+            updatedOwner,
+            selectedPlan,
+            ownerPreviousPlan,
+            ownerProperties,
+            propsToDeactivate,
+            updatedOwner.paymentData,
+            session,
+          )
         }
       } else {
         throw new BadRequestException(
@@ -1141,7 +1117,7 @@ export class CreateProperty_Service {
         creditCard,
         ownerActiveProperties,
         propertiesToDeactivate,
-        session
+        session,
       } = planTransitionDto
 
       let updatedOwner: any = owner
@@ -1178,8 +1154,8 @@ export class CreateProperty_Service {
         ownerActiveProperties,
         propertiesToDeactivate,
         updatedOwner.paymentData,
-        session
-      );
+        session,
+      )
 
       return updatedOwner
     } catch (error) {
@@ -1310,84 +1286,89 @@ export class CreateProperty_Service {
     ownerProperties: any,
     propsToDeactivate: any,
     payment: any,
-    session: any
+    session: any,
   ): Promise<any> {
     try {
-      const { paymentData, _id } = owner;
-      let ownerId;
+      const { _id } = owner
+      let ownerId = _id
 
-      let updatedOwner = owner;
+      let updatedOwner = owner
       let newAdCredits
       const newHighlightCredits = owner.highlightCredits + newPlan.highlightAd
 
       // Verificar se apesar do plano ser grátis o owner ainda possuia créditos de outro plano anterior;
-      if (previousPlan && previousPlan.toString() !== newPlan._id.toString()) {
+      if (
+        previousPlan &&
+        previousPlan._id.toString() !== newPlan._id.toString()
+      ) {
         newAdCredits = owner.adCredits + newPlan.commonAd
         newAdCredits =
           newAdCredits - (ownerProperties.length - propsToDeactivate.length)
-      } else {
-        newAdCredits = owner.adCredits
-      }
 
-      if (payment?.subscriptionId) {
-        // Buscar data de vencimento;
-        const { data } = await axios.get(
-          `${process.env.PAYMENT_URL}/payment/subscription/${payment?.subscriptionId}`,
-          {
-            headers: {
-              'Content-Type': 'application/json',
-              access_token: process.env.ASAAS_API_KEY || '',
-            },
-            timeout: 3000000,
-          },
-        )
-
-        const updateDate = new Date(data.nextDueDate)
-
-        // Lógica de mudança de créditos imediata;
-        updatedOwner = {
-          ...owner,
-          adCredits: newAdCredits - 1,
-          highlightCredits: newHighlightCredits,
-          planTransitionStatus: 'processing',
-        }
-
-        if (!ownerId) {
-          const createdOwner = await this.ownerModel.create([updatedOwner], {
-            session,
-          })
-          ownerId = createdOwner[0]._id
-          updatedOwner = createdOwner[0]
-        }
-        // Verificar se já existe um cron job com este ID
-        if (this.schedulerRegistry.doesExist('cron', _id.toString())) {
-          // Se existe, deletar o cron job atual
-          this.schedulerRegistry.deleteCronJob(_id.toString())
-          this.logger.debug(`Job anterior para o anunciante ${_id} deletado`)
-        }
-
-        // Criar novo cron job
-        const job = new CronJob(updateDate, async () => {
-          await this.ownerModel.updateOne(
-            { _id },
+        if (payment?.subscriptionId) {
+          // Buscar data de vencimento;
+          const { data } = await axios.get(
+            `${process.env.PAYMENT_URL}/payment/subscription/${payment?.subscriptionId}`,
             {
-              $set: {
-                adCredits: newAdCredits - owner.adCredits,
-                highlightCredits: newHighlightCredits - owner.highlightCredits,
-                planTransitionStatus: 'complete',
+              headers: {
+                'Content-Type': 'application/json',
+                access_token: process.env.ASAAS_API_KEY || '',
               },
+              timeout: 3000000,
             },
           )
-          this.logger.debug(`Anunciante ${_id} atualizado`)
-          this.schedulerRegistry.deleteCronJob(_id.toString()) // remove o job após execução
-        })
 
-        this.schedulerRegistry.addCronJob(_id.toString(), job)
-        job.start()
-        this.logger.debug(
-          `Job para o anunciante ${_id} agendado para: ${updateDate}`,
-        )
+          const updateDate = new Date(data.nextDueDate)
+
+          // Lógica de mudança de créditos imediata;
+          updatedOwner = {
+            ...owner,
+            adCredits: newAdCredits,
+            highlightCredits: newHighlightCredits,
+            planTransitionStatus: 'processing',
+          }
+
+          if (!ownerId) {
+            const createdOwner = await this.ownerModel.create([updatedOwner], {
+              session,
+            })
+            ownerId = createdOwner[0]._id
+            updatedOwner = createdOwner[0]
+          }
+          // Verificar se já existe um cron job com este ID
+          if (this.schedulerRegistry.doesExist('cron', _id.toString())) {
+            // Se existe, deletar o cron job atual
+            this.schedulerRegistry.deleteCronJob(_id.toString())
+            this.logger.debug(`Job anterior para o anunciante ${_id} deletado`)
+          }
+
+          // Criar novo cron job
+          const job = new CronJob(updateDate, async () => {
+            await this.ownerModel.updateOne(
+              { _id },
+              {
+                $set: {
+                  adCredits: newAdCredits - owner.adCredits,
+                  highlightCredits:
+                    newHighlightCredits - owner.highlightCredits,
+                  planTransitionStatus: 'complete',
+                },
+              },
+            )
+            this.logger.debug(`Anunciante ${_id} atualizado`)
+            this.schedulerRegistry.deleteCronJob(_id.toString()) // remove o job após execução
+          })
+
+          this.schedulerRegistry.addCronJob(_id.toString(), job)
+          job.start()
+          this.logger.debug(
+            `Job para o anunciante ${_id} agendado para: ${updateDate}`,
+          )
+        } else {
+          updatedOwner.adCredits = newAdCredits - 1
+        }
       } else {
+        newAdCredits = owner.adCredits - 1
         updatedOwner.adCredits = newAdCredits
       }
